# Modifications for small inputs (xxh3)