--- a/xxh3.h
+++ b/xxh3.h
@@ -532,7 +532,7 @@ XXH3_mul128_fold64(xxh_u64 lhs, xxh_u64 rhs)
 static XXH64_hash_t XXH3_avalanche(xxh_u64 h64)
 {
     h64 ^= h64 >> 37;
-    h64 *= PRIME64_3;
+    h64 *= 0x165667919E3779F9ULL;
     h64 ^= h64 >> 32;
     return h64;
 }
@@ -551,7 +551,7 @@ XXH3_len_1to3_64b(const xxh_u8* input, size_t len, const xxh_u8* secret, XXH64_h
     {   xxh_u8 const c1 = input[0];
         xxh_u8 const c2 = input[len >> 1];
         xxh_u8 const c3 = input[len - 1];
-        xxh_u32  const combined = ((xxh_u32)c1) | (((xxh_u32)c2) << 8) | (((xxh_u32)c3) << 16) | (((xxh_u32)len) << 24);
+        xxh_u32  const combined = ((xxh_u32)c1<<16) | (((xxh_u32)c2) << 24) | (((xxh_u32)c3) << 0) | (((xxh_u32)len) << 8);
         xxh_u64  const keyed = (xxh_u64)combined ^ (XXH_readLE32(secret) + seed);
         xxh_u64  const mixed = keyed * PRIME64_1;
         return XXH3_avalanche(mixed);
@@ -564,12 +564,12 @@ XXH3_len_4to8_64b(const xxh_u8* input, size_t len, const xxh_u8* secret, XXH64_h
     XXH_ASSERT(input != NULL);
     XXH_ASSERT(secret != NULL);
     XXH_ASSERT(4 <= len && len <= 8);
-    {   xxh_u32 const input_lo = XXH_readLE32(input);
-        xxh_u32 const input_hi = XXH_readLE32(input + len - 4);
-        xxh_u64 const input_64 = input_lo | ((xxh_u64)input_hi << 32);
-        xxh_u64 const keyed = input_64 ^ (XXH_readLE64(secret) + seed);
-        xxh_u64 const mix64 = len + ((keyed ^ (keyed >> 51)) * PRIME32_1);
-        return XXH3_avalanche((mix64 ^ (mix64 >> 47)) * PRIME64_2);
+    {   xxh_u32 const input_1 = XXH_readLE32(input);
+        xxh_u32 const input_2 = XXH_readLE32(input + len - 4);
+        xxh_u64 const input_64 = input_2 | ((xxh_u64)input_1 << 32);
+        xxh_u64 const keyed = input_64 ^ (seed + XXH_readLE64(secret));
+        xxh_u64 const mix64 = len + ((keyed ^ (keyed >> 39)) * PRIME64_2);
+        return XXH3_avalanche(mix64 ^ (mix64 >> 47));
     }
 }
 
@@ -579,7 +579,7 @@ XXH3_len_9to16_64b(const xxh_u8* input, size_t len, const xxh_u8* secret, XXH64_
     XXH_ASSERT(input != NULL);
     XXH_ASSERT(secret != NULL);
     XXH_ASSERT(9 <= len && len <= 16);
-    {   xxh_u64 const input_lo = XXH_readLE64(input)           ^ (XXH_readLE64(secret)     + seed);
+    {   xxh_u64 const input_lo = XXH_readLE64(input)           ^  XXH_readLE64(secret);
         xxh_u64 const input_hi = XXH_readLE64(input + len - 8) ^ (XXH_readLE64(secret + 8) - seed);
         xxh_u64 const acc = len + (input_lo + input_hi) + XXH3_mul128_fold64(input_lo, input_hi);
         return XXH3_avalanche(acc);
@@ -593,7 +593,7 @@ XXH3_len_0to16_64b(const xxh_u8* input, size_t len, const xxh_u8* secret, XXH64_
     {   if (len > 8) return XXH3_len_9to16_64b(input, len, secret, seed);
         if (len >= 4) return XXH3_len_4to8_64b(input, len, secret, seed);
         if (len) return XXH3_len_1to3_64b(input, len, secret, seed);
-        return 0;
+        return XXH3_avalanche((PRIME64_1 + seed) ^ XXH_readLE64(secret));
     }
 }
 
@@ -1426,7 +1426,9 @@ XXH3_len_0to16_128b(const xxh_u8* input, size_t len, const xxh_u8* secret, XXH64
     {   if (len > 8) return XXH3_len_9to16_128b(input, len, secret, seed);
         if (len >= 4) return XXH3_len_4to8_128b(input, len, secret, seed);
         if (len) return XXH3_len_1to3_128b(input, len, secret, seed);
-        {   XXH128_hash_t const h128 = { 0, 0 };
+        {   XXH128_hash_t h128;
+            h128.low64 = XXH3_avalanche((PRIME64_1 + seed) ^ XXH_readLE64(secret));
+            h128.high64 = XXH3_avalanche((PRIME64_2 - seed) ^ XXH_readLE64(secret+8));
             return h128;
     }   }
 }