The modifications pass the new Perlin noise test (not yet published).
It also helps speed for the 4-8 range (latency test from 112MH/s to 125MH/s).

I think the speed could be improved even more, but to do this, it would be necessary to drop the "64-bit bijective" property (0-collisions when all inputs are 64-bit). I haven't ruled that yet, but if bijection is not important it could be the next move.

I will have to port the modifications to `xxh128` too.

The final avalanche has been slightly altered, improving bit distribution bias across the range.

This however changes all values produced by `xxh3`. Some verification values in `xxhsum` still need updating.