@easyaspi314 2020-02-14
XXH128 does have a lot of latency on short keys, it would be nice to see some improvements.

For keys <= 16 bytes:

32-bit: XXH32 > XXH3_64b >>> XXH64 > XXH128
x86_64: XXH32 ≈ XXH3_64b > XXH128 > XXH64
AArch64: XXH3_64b > XXH32 > XXH128 > XXH64

XXH3 is faster than XXH32 on AArch64 mainly because it can xor fold without any temporaries:
```asm
        eor     x0, x0, x0, lsr #47
```
compared to x86_64:
```asm
        mov     rcx, rax
        shr     rcx, 47
        xor     rax, rcx
```