@Cyan4973 2020-02-14
Dropping the requirement of bijectivity for 64-bit inputs, the speed climbs to > 150 MH/s for the 4-7 range, and > 160 MH/s for the length 8 itself.
So it's pretty big boost.

Apart from the loss of bijectivity, all other quality metrics seem preserved (good dispersion, collisions in accordance with birthday paradox).