@easyaspi314 2020-02-14
What does the non-bijective version look like?