@easyaspi314 2020-02-17
On my Pixel:

64-bit
```
./xxhsum 0.7.3 (64-bits aarch64 little endian), Clang 9.0.1 , by Yann Collet
Sample of 5 bytes...
XXH32                    :          5 -> 78979139 it/s (  376.6 MB/s)
XXH32 unaligned          :          5 -> 80623227 it/s (  384.4 MB/s)
XXH64                    :          5 -> 52919520 it/s (  252.3 MB/s)
XXH64 unaligned          :          5 -> 50402130 it/s (  240.3 MB/s)
XXH3_64b                 :          5 -> 99918214 it/s (  476.4 MB/s)
XXH3_64b unaligned       :          5 -> 100049695 it/s (  477.1 MB/s)
XXH3_64b seeded          :          5 -> 94150987 it/s (  448.9 MB/s)
XXH3_64b seeded unaligne :          5 -> 94381426 it/s (  450.0 MB/s)
XXH128                   :          5 -> 66374743 it/s (  316.5 MB/s)
XXH128 unaligned         :          5 -> 66631507 it/s (  317.7 MB/s)
XXH128 seeded            :          5 -> 61666322 it/s (  294.0 MB/s)
XXH128 seeded unaligned  :          5 -> 61957284 it/s (  295.4 MB/s)
XXH3_new                 :          5 -> 104639925 it/s (  499.0 MB/s)
XXH3_new unaligned       :          5 -> 105611273 it/s (  503.6 MB/s)
XXH3_new seeded          :          5 -> 100946493 it/s (  481.4 MB/s)
XXH3_new seeded unaligne :          5 -> 100562905 it/s (  479.5 MB/s)
```
```
clang -O3 -m32 -march=armv6t2
$ ./xxhsum -b -B5
Sample of 5 bytes...
XXH32                    :          5 -> 57626760 it/s (  274.8 MB/s)
XXH32 unaligned          :          5 -> 59118209 it/s (  281.9 MB/s)
XXH64                    :          5 -> 27138541 it/s (  129.4 MB/s)
XXH64 unaligned          :          5 -> 25513446 it/s (  121.7 MB/s)
XXH3_64b                 :          5 -> 45701062 it/s (  217.9 MB/s)
XXH3_64b unaligned       :          5 -> 45433547 it/s (  216.6 MB/s)
XXH3_64b seeded          :          5 -> 43698427 it/s (  208.4 MB/s)
XXH3_64b seeded unaligne :          5 -> 43364902 it/s (  206.8 MB/s)
XXH128                   :          5 -> 27789511 it/s (  132.5 MB/s)
XXH128 unaligned         :          5 -> 27637196 it/s (  131.8 MB/s)
XXH128 seeded            :          5 -> 25321305 it/s (  120.7 MB/s)
XXH128 seeded unaligned  :          5 -> 25297310 it/s (  120.6 MB/s)
XXH3_new                 :          5 -> 50940968 it/s (  242.9 MB/s)
XXH3_new unaligned       :          5 -> 50938749 it/s (  242.9 MB/s)
XXH3_new seeded          :          5 -> 55158194 it/s (  263.0 MB/s)
XXH3_new seeded unaligne :          5 -> 54682827 it/s (  260.7 MB/s)
```

It's interesting how the seeded variant is faster. I'm guessing it takes advantage of a barrel shift or a multiply+accumulate.