@easyaspi314 2020-02-18
Oops, just realized I forgot `-munaligned-access -fomit-frame-pointer` which I usually use.

Here is `clang -O3 -m32 -march=armv6t2 -munaligned-access -fomit-frame-pointer`

```
./xxhsum 0.7.3 (32-bits arm little endian), Clang 9.0.1 , by Yann Collet
Sample of 5 bytes...
XXH32                    :          5 -> 73493698 it/s (  350.4 MB/s)
XXH32 unaligned          :          5 -> 71677171 it/s (  341.8 MB/s)
XXH64                    :          5 -> 29337822 it/s (  139.9 MB/s)
XXH64 unaligned          :          5 -> 29028419 it/s (  138.4 MB/s)
XXH3_64b                 :          5 -> 52521477 it/s (  250.4 MB/s)
XXH3_64b unaligned       :          5 -> 52145183 it/s (  248.6 MB/s)
XXH3_64b seeded          :          5 -> 54742558 it/s (  261.0 MB/s)
XXH3_64b seeded unaligne :          5 -> 54912692 it/s (  261.8 MB/s)
XXH128                   :          5 -> 31115296 it/s (  148.4 MB/s)
XXH128 unaligned         :          5 -> 31410234 it/s (  149.8 MB/s)
XXH128 seeded            :          5 -> 29905792 it/s (  142.6 MB/s)
XXH128 seeded unaligned  :          5 -> 29889476 it/s (  142.5 MB/s)
XXH3_new                 :          5 -> 73039944 it/s (  348.3 MB/s)
XXH3_new unaligned       :          5 -> 72427877 it/s (  345.4 MB/s)
XXH3_new seeded          :          5 -> 66844700 it/s (  318.7 MB/s)
XXH3_new seeded unaligne :          5 -> 68765537 it/s (  327.9 MB/s)
```

That benchmark better matches the AArch64 version, as they shouldn't be too far off thanks ARM's fast 64-bit multiplies.