@Cyan4973 2020-02-18
> What does the non-bijective version look like?

There are multiple versions possible, but I would start by one resembling the 8-16 segment, which has already proven its worth in term of bit dispersion.
I will make it a separate branch, so that it can be compared with the current one.

Thanks for your measurements on ARM, it's very useful. It shows that this variant is already a sensible improvement for ARM. It will be useful again when comparing with the non-bijective variant.