@easyaspi314 2020-02-18
No problem.

As for bijectivity, as long as it doesn't have any major collisions, I don't think it is necessary.

Besides, in practice, inputs will be all different lengths.