@easyaspi314 2020-02-19
As for performance of the 9-16 variant:
 - 64-bit is, unsurprisingly, no contest.
 - 32-bit is faster than XXH32 on lengths 11 and 13-16.

Of course, ARM is naturally going to be better at the 9-16 routine than x86, as the long multiply is only 4 (8 without `umaal`) instructions compared to the 20-30 required on x86.