@easyaspi314 2020-02-19
If it helps, here is the output of `./benchHash --maxl=0 --maxs=16`. 
`clang -O3`
[aarch64.txt](https://github.com/Cyan4973/xxHash/files/4222802/aarch64.txt)
`clang -O3 -m32 -march=armv6t2 -munaligned-access -fomit-frame-pointer`
[armv6t2.txt](https://github.com/Cyan4973/xxHash/files/4222817/armv6t2.txt)