@easyaspi314 2020-02-19
[I made spreadsheets/charts of the latency for aarch64, arm, x86, and x86_64](https://docs.google.com/spreadsheets/d/1Gb6SFF2gsndmelb_Lq2mBaaTJ1Z-pb6_qhaAn3aW1ak/edit?usp=sharing)

I realized that Clang is vectorizing the XXH128 short inputs on i686, significantly reducing the speed. I wonder if I can fix this.