@easyaspi314 2020-02-19
```
Build started
git clone -q --depth=2 --branch=smallFold https://github.com/Cyan4973/xxHash.git C:\projects\xxhash
git checkout -qf 6456c04490667acf035884e91aac32208d09dbbe
fatal: reference is not a tree: 6456c04490667acf035884e91aac32208d09dbbe
Command exited with code 128
```
Not sure what that is about.