@easyaspi314 2020-02-19
Added the x86 benchmarks to the spreadsheet for your `smallFold` branch I noticed, will update ARM benches later.