@Cyan4973 2020-02-19
Thanks for benchmarking @easyaspi314 ,
from these measurements, it seems that `smallFold` is beneficial for `x64`, but detrimental for `x86`.

Note that also measured speed on a local `aarch64` system,
and `smallFold` seems detrimental on this system too (compared to `smallInputs`).