@easyaspi314 2020-02-20
That is to be expected.

aarch64 has a significant overhead on 64-bit and 64->128-bit multiplies.

Unlike most x86_64 chips which have basically the same speed, most aarch64 chips have a 5 cycle `madd/mul` (x-form) and a 6 cycle `umulh`, and some chips even stall the pipeline a bit. This means that `XXH_mult64to128` takes roughly 11 cycles on average, compared to x86_64's 2-4 cycles.

That is also the reason why XXH64 is roughly the same speed as XXH32: The speedup from increased bandwidth is cancelled out by the slower `mul`.